Refactor Vagrantfile to support more varied vm cfg
Move the common configuration into a configure_vm method that takes in
the specific vm config and an options hash. This method should be called
inside of a 'config.vm.define' block, permitting VM's to set the
'use_nfs' option to true if they wish to be configured with NFS instead
of vboxfs.

Also hack the Vagrant::Guest::FreeBSD class to include the fix for
https://github.com/mitchellh/vagrant/issues/1321 because omnibus-chef
uses the older vagrant 1.0.x series.

Use pathname to expand paths completely for mounts.